+++ truffle/src/com.oracle.truffle.api.test.jdk17/src/com/oracle/truffle/api/test/polyglot/RegisterJDK17TestClassesForReflectionFeature.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -42,39 +42,14 @@ package com.oracle.truffle.api.test.polyglot;
 
 import java.util.List;
 
-import org.graalvm.nativeimage.hosted.Feature;
-import org.graalvm.nativeimage.hosted.RuntimeReflection;
-
-import com.oracle.truffle.tck.tests.ValueAssert;
-
 /**
  * Register all <i>named</i> inner classes under {@link #TEST_CLASSES} for reflection, so that they
  * do not need to be listed in {@code reflection.json}. Anonymous inner classes however must be
  * listed in {@code reflection.json}, as Java provides no way to list those.
  */
-public class RegisterTestClassesForReflectionFeature implements Feature {
+public class RegisterJDK17TestClassesForReflectionFeature extends RegisterTestClassesForReflectionFeature {
 
-    private static final List<Class<?>> TEST_CLASSES = List.of(
-                    ValueHostInteropTest.class,
-                    ValueHostConversionTest.class,
-                    ValueAssert.class,
-                    ValueAPITest.class,
-                    ValueScopingTest.class,
-                    PolyglotExceptionTest.class,
-                    LanguageSPIHostInteropTest.class,
-                    HostAccessTest.class,
-                    ExposeToGuestTest.class,
-                    ContextAPITest.class);
-
-    protected static void registerClass(Class<?> clazz) {
-        RuntimeReflection.register(clazz);
-        RuntimeReflection.register(clazz.getConstructors());
-        RuntimeReflection.register(clazz.getDeclaredConstructors());
-        RuntimeReflection.register(clazz.getMethods());
-        RuntimeReflection.register(clazz.getDeclaredMethods());
-        RuntimeReflection.register(clazz.getFields());
-        RuntimeReflection.register(clazz.getDeclaredFields());
-    }
+    private static final List<Class<?>> TEST_CLASSES = List.of(HostRecordAccessTest.class);
 
     public void beforeAnalysis(BeforeAnalysisAccess access) {
         for (Class<?> testClass : TEST_CLASSES) {