+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ContextAPITest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -118,7 +118,10 @@ public class ContextAPITest extends AbstractPolyglotTest {
 
     @BeforeClass
     public static void initHostAccess() throws Exception {
-        CONFIG = HostAccess.newBuilder().allowAccess(Runnable.class.getMethod("run")).allowAccessAnnotatedBy(HostAccess.Export.class).build();
+        CONFIG = HostAccess.newBuilder().//
+                        allowAccess(Runnable.class.getMethod("run")).//
+                        allowAccessAnnotatedBy(HostAccess.Export.class).//
+                        allowAccessInheritance(true).build();
     }
 
     public ContextAPITest() {